Flattened extensions to LightBDD.Framework.Scenarios
