+++ src/LightBDD.Framework/Scenarios/CompactExtensions.cs
@@ -1,16 +1,15 @@
-﻿using LightBDD.Framework.Extensibility;
-using LightBDD.Framework.Scenarios.Compact.Implementation;
-using System;
+﻿using System;
 using System.Diagnostics;
 using System.Threading.Tasks;
+using LightBDD.Framework.Scenarios.Implementation;
 
-namespace LightBDD.Framework.Scenarios.Compact
+namespace LightBDD.Framework.Scenarios
 {
     /// <summary>
     /// Extensions class allowing to use compact steps for defining LightBDD test steps.
     /// </summary>
     [DebuggerStepThrough]
-    public static class CompactStepExtensions
+    public static class CompactExtensions
     {
         /// <summary>
         /// Adds step with name specified by <paramref name="name"/> parameter and action specified by <paramref name="step"/> parameter to the composite step.
@@ -38,6 +37,36 @@ namespace LightBDD.Framework.Scenarios.Compact
         {
             builder.Integrate().AddSteps(new[] { CompactStepCompiler.ToAsynchronousStep(name, step) });
             return builder;
+        }
+
+        /// <summary>
+        /// Adds step with name specified by <paramref name="name"/> parameter and action specified by <paramref name="step"/> parameter to the scenario.
+        /// </summary>
+        /// <typeparam name="TContext">Scenario context, if specified.</typeparam>
+        /// <param name="builder">Builder.</param>
+        /// <param name="name">Step name to be rendered.</param>
+        /// <param name="step">Step action</param>
+        /// <returns>Builder.</returns>
+        public static IScenarioRunner<TContext> AddStep<TContext>(this IScenarioBuilder<TContext> builder, string name, Action<TContext> step)
+        {
+            var integration = builder.Integrate();
+            integration.Core.AddSteps(new[] { CompactStepCompiler.ToSynchronousStep(name, step) });
+            return integration;
+        }
+
+        /// <summary>
+        /// Adds asynchronous step with name specified by <paramref name="name"/> parameter and action specified by <paramref name="step"/> parameter to the scenario.
+        /// </summary>
+        /// <typeparam name="TContext">Scenario context, if specified.</typeparam>
+        /// <param name="builder">Builder.</param>
+        /// <param name="name">Step name to be rendered.</param>
+        /// <param name="step">Step action</param>
+        /// <returns>Builder.</returns>
+        public static IScenarioRunner<TContext> AddAsyncStep<TContext>(this IScenarioBuilder<TContext> builder, string name, Func<TContext, Task> step)
+        {
+            var integration = builder.Integrate();
+            integration.Core.AddSteps(new[] { CompactStepCompiler.ToAsynchronousStep(name, step) });
+            return integration;
         }
     }
 }
