+++ test/LightBDD.Fixie2.UnitTests/IntegrationTests.cs
@@ -3,8 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using LightBDD.Core.Results;
 using LightBDD.Framework;
-using LightBDD.Framework.Scenarios.Basic;
-using LightBDD.Framework.Scenarios.Extended;
+using LightBDD.Framework.Scenarios;
 using Shouldly;
 
 namespace LightBDD.Fixie2.UnitTests
