+++ test/LightBDD.Framework.UnitTests/Scenarios/Basic/Basic_scenario_exception_stack_trace_integration_tests.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using LightBDD.Framework.Extensibility;
-using LightBDD.Framework.Scenarios.Basic;
+using LightBDD.Framework.Scenarios;
 using LightBDD.UnitTests.Helpers;
 using LightBDD.UnitTests.Helpers.TestableIntegration;
 using NUnit.Framework;
@@ -26,9 +26,9 @@ namespace LightBDD.Framework.UnitTests.Scenarios.Basic
             var ex = Assert.Throws<InvalidOperationException>(() => _runner.RunScenario(Step_throwing_exception));
             ex.AssertStackTraceMatching(
 @"^\s*at LightBDD.Framework.UnitTests.Scenarios.Basic.Basic_scenario_exception_stack_trace_integration_tests.Step_throwing_exception[^\n]*
-\s*at LightBDD.Framework.Scenarios.Basic.Implementation.BasicStepCompiler.StepExecutor.Execute[^\n]*
+\s*at LightBDD.Framework.Scenarios.Implementation.BasicStepCompiler.StepExecutor.Execute[^\n]*
 ([^\n]*
-)?\s*at LightBDD.Framework.Scenarios.Basic.BasicScenarioExtensions.RunScenario[^\n]*");
+)?\s*at LightBDD.Framework.Scenarios.BasicExtensions.RunScenario[^\n]*");
         }
 
         void Step_throwing_exception()
@@ -43,7 +43,7 @@ namespace LightBDD.Framework.UnitTests.Scenarios.Basic
             ex.AssertStackTraceMatching(
                 @"^\s*at LightBDD.Framework.UnitTests.Scenarios.Basic.Basic_scenario_exception_stack_trace_integration_tests[^\n]*Async_step_throwing_exception_immediately[^\n]*
 ([^\n]*
-)?\s*at LightBDD.Framework.Scenarios.Basic.BasicScenarioExtensions[^\n]*RunScenarioAsync[^\n]*");
+)?\s*at LightBDD.Framework.Scenarios.BasicExtensions[^\n]*RunScenarioAsync[^\n]*");
         }
 
         async Task Async_step_throwing_exception_immediately()
@@ -58,7 +58,7 @@ namespace LightBDD.Framework.UnitTests.Scenarios.Basic
             ex.AssertStackTraceMatching(
                 @"^\s*at LightBDD.Framework.UnitTests.Scenarios.Basic.Basic_scenario_exception_stack_trace_integration_tests[^\n]*Async_step_throwing_exception_after_await[^\n]*
 ([^\n]*
-)?\s*at LightBDD.Framework.Scenarios.Basic.BasicScenarioExtensions[^\n]*RunScenarioAsync[^\n]*");
+)?\s*at LightBDD.Framework.Scenarios.BasicExtensions[^\n]*RunScenarioAsync[^\n]*");
         }
 
         async Task Async_step_throwing_exception_after_await()
