+++ test/LightBDD.Framework.UnitTests/Scenarios/Basic/Basic_scenario_runner_hierarchical_execution_tests.cs
@@ -1,9 +1,8 @@
 using System.Linq;
 using System.Threading.Tasks;
 using LightBDD.Core.Extensibility.Results;
-using LightBDD.Framework.Scenarios.Basic;
+using LightBDD.Framework.Scenarios;
 using LightBDD.Framework.UnitTests.Scenarios.Basic.Helpers;
-using LightBDD.Framework.UnitTests.Scenarios.Helpers;
 using NUnit.Framework;
 
 #pragma warning disable CS1998 // Async method lacks 'await' operators and will run synchronously
