+++ test/LightBDD.Framework.UnitTests/Scenarios/Contextual/Contextual_runner_context_disposal_tests.cs
@@ -1,11 +1,11 @@
 ﻿using LightBDD.Core.Dependencies;
 using LightBDD.Core.Extensibility;
-using LightBDD.Framework.Scenarios.Contextual;
 using LightBDD.Framework.UnitTests.Scenarios.Helpers;
 using Moq;
 using NUnit.Framework;
 using System;
 using System.Collections.Generic;
+using LightBDD.Framework.Scenarios;
 
 namespace LightBDD.Framework.UnitTests.Scenarios.Contextual
 {
