+++ test/LightBDD.Framework.UnitTests/Scenarios/Contextual/Contextual_runner_hierarchical_execution_tests.cs
@@ -1,8 +1,6 @@
 ﻿using System;
 using LightBDD.Core.Configuration;
 using LightBDD.Framework.Scenarios;
-using LightBDD.Framework.Scenarios.Basic;
-using LightBDD.Framework.Scenarios.Contextual;
 using LightBDD.Framework.UnitTests.Scenarios.Extended.Helpers;
 using LightBDD.Framework.UnitTests.Scenarios.Helpers;
 using NUnit.Framework;
