+++ test/LightBDD.Framework.UnitTests/Scenarios/Extended/Extended_scenario_runner_execution_tests.cs
@@ -1,11 +1,11 @@
 using LightBDD.Core.Execution;
-using LightBDD.Framework.Scenarios.Extended;
 using LightBDD.Framework.UnitTests.Scenarios.Extended.Helpers;
 using LightBDD.Framework.UnitTests.Scenarios.Helpers;
 using NUnit.Framework;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
+using LightBDD.Framework.Scenarios;
 
 namespace LightBDD.Framework.UnitTests.Scenarios.Extended
 {
