+++ test/LightBDD.Framework.UnitTests/Scenarios/Extended/Extended_scenario_runner_hierarchical_execution_tests.cs
@@ -1,11 +1,11 @@
 using LightBDD.Core.Extensibility.Results;
-using LightBDD.Framework.Scenarios.Extended;
 using LightBDD.Framework.UnitTests.Scenarios.Extended.Helpers;
 using LightBDD.Framework.UnitTests.Scenarios.Helpers;
 using NUnit.Framework;
 using System;
 using System.Linq;
 using System.Threading.Tasks;
+using LightBDD.Framework.Scenarios;
 
 #pragma warning disable CS1998 // Async method lacks 'await' operators and will run synchronously
