+++ test/LightBDD.Framework.UnitTests/Scenarios/Fluent/Fluent_scenario_runner_execution_tests.cs
@@ -1,11 +1,9 @@
 using LightBDD.Framework.Extensibility;
-using LightBDD.Framework.Scenarios.Basic;
-using LightBDD.Framework.Scenarios.Contextual;
-using LightBDD.Framework.Scenarios.Extended;
 using LightBDD.Framework.UnitTests.Scenarios.Extended.Helpers;
 using LightBDD.UnitTests.Helpers.TestableIntegration;
 using NUnit.Framework;
 using System;
+using LightBDD.Framework.Scenarios;
 
 namespace LightBDD.Framework.UnitTests.Scenarios.Fluent
 {
