+++ test/LightBDD.MsTest2.UnitTests/Cross_domain_tests.cs
@@ -2,7 +2,7 @@
 using System.Configuration;
 using System.Diagnostics;
 using LightBDD.Framework;
-using LightBDD.Framework.Scenarios.Extended;
+using LightBDD.Framework.Scenarios;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace LightBDD.MsTest2.UnitTests
