+++ test/LightBDD.NUnit2.NUnitTests/Cross_domain_tests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Configuration;
 using System.Diagnostics;
 using LightBDD.Framework;
-using LightBDD.Framework.Scenarios.Extended;
+using LightBDD.Framework.Scenarios;
 using NUnit.Framework;
 
 namespace LightBDD.NUnit2.NUnitTests
