+++ test/LightBDD.UnitTests.Helpers/TestableIntegration/TestSyntaxRunner.cs
@@ -5,8 +5,6 @@ using System.Threading.Tasks;
 using LightBDD.Core.Execution;
 using LightBDD.Core.Extensibility;
 using LightBDD.Framework;
-using LightBDD.Framework.Scenarios;
-using LightBDD.Framework.Scenarios.Contextual;
 
 namespace LightBDD.UnitTests.Helpers.TestableIntegration
 {
