+++ test/LightBDD.UnitTests.Helpers/TestableIntegration/TestSyntaxRunnerExtensions.cs
@@ -1,5 +1,4 @@
 using LightBDD.Framework;
-using LightBDD.Framework.Extensibility;
 
 namespace LightBDD.UnitTests.Helpers.TestableIntegration
 {
